Added arrange act assert. Migrated sign in method to other project sign in tests.
