+++ tests/Api.Integration.Tests/Endpoints/SignIn/SignInTests.cs
@@ -41,10 +41,13 @@ public class SignInTests : IClassFixture<PasswordlessApiFactory>
     [Fact]
     public async Task I_can_retrieve_assertion_options_to_begin_sign_in()
     {
+        // Arrange
         var request = new SignInBeginDTO { Origin = OriginUrl, RPID = RpId };
 
+        // Act
         var response = await _httpClient.PostAsJsonAsync("/signin/begin", request);
 
+        // Assert
         response.StatusCode.Should().Be(HttpStatusCode.OK);
         var signInResponse = await response.Content.ReadFromJsonAsync<SessionResponse<Fido2NetLib.AssertionOptions>>();
 
@@ -57,6 +60,7 @@ public class SignInTests : IClassFixture<PasswordlessApiFactory>
     [Fact]
     public async Task I_can_retrieve_my_passkey_after_registering_and_receive_a_sign_in_token()
     {
+        // Arrange
         var tokenRequest = TokenGenerator.Generate();
         var tokenResponse = await _httpClient.PostAsJsonAsync("/register/token", tokenRequest);
         var registerTokenResponse = await tokenResponse.Content.ReadFromJsonAsync<RegisterEndpoints.RegisterTokenResponse>();
@@ -91,6 +95,7 @@ public class SignInTests : IClassFixture<PasswordlessApiFactory>
                                                 $"return await getCredential({signInBegin!.Data.ToJson()});")?.ToString() ?? string.Empty;
         var parsedSignInResult = JsonSerializer.Deserialize<AuthenticatorAssertionRawResponse>(signInResult);
 
+        // Act
         var signInCompleteResponse = await _httpClient.PostAsJsonAsync("/signin/complete", new SignInCompleteDTO
         {
             Origin = OriginUrl,
@@ -99,6 +104,7 @@ public class SignInTests : IClassFixture<PasswordlessApiFactory>
             Session = signInBegin.Session
         });
 
+        // Assert
         signInCompleteResponse.StatusCode.Should().Be(HttpStatusCode.OK);
         var signInTokenResponse = await signInCompleteResponse.Content.ReadFromJsonAsync<TokenResponse>();
         signInTokenResponse.Should().NotBeNull();
@@ -108,6 +114,7 @@ public class SignInTests : IClassFixture<PasswordlessApiFactory>
     [Fact]
     public async Task I_can_retrieve_my_passkey_after_registering_and_receive_a_valid_sign_in_token()
     {
+        // Arrange
         var tokenRequest = TokenGenerator.Generate();
         var tokenResponse = await _httpClient.PostAsJsonAsync("/register/token", tokenRequest);
         var registerTokenResponse = await tokenResponse.Content.ReadFromJsonAsync<RegisterEndpoints.RegisterTokenResponse>();
@@ -143,6 +150,7 @@ public class SignInTests : IClassFixture<PasswordlessApiFactory>
                                                 $"return await getCredential({signInBegin!.Data.ToJson()});")?.ToString() ?? string.Empty;
         var parsedSignInResult = JsonSerializer.Deserialize<AuthenticatorAssertionRawResponse>(signInResult);
 
+        // Act
         var signInCompleteResponse = await _httpClient.PostAsJsonAsync("/signin/complete", new SignInCompleteDTO
         {
             Origin = OriginUrl,
@@ -151,6 +159,7 @@ public class SignInTests : IClassFixture<PasswordlessApiFactory>
             Session = signInBegin.Session
         });
 
+        // Assert
         signInCompleteResponse.StatusCode.Should().Be(HttpStatusCode.OK);
         var signInTokenResponse = await signInCompleteResponse.Content.ReadFromJsonAsync<TokenResponse>();
         signInTokenResponse.Should().NotBeNull();
@@ -158,5 +167,48 @@ public class SignInTests : IClassFixture<PasswordlessApiFactory>
 
         var verifySignInResponse = await _httpClient.PostAsJsonAsync("/signin/verify", new SignInVerifyDTO { Token = signInTokenResponse.Token });
         verifySignInResponse.StatusCode.Should().Be(HttpStatusCode.OK);
+    }
+    
+    [Fact]
+    public async Task UnknownCredentialThrows()
+    {
+        var options = await _httpClient.PostAsJsonAsync("/signin/begin", new { Origin = "https://localhost", RPID = "localhost" });
+
+        var response = await options.Content.ReadFromJsonAsync<SessionResponse<Fido2NetLib.AssertionOptions>>();
+
+        var payload = new
+        {
+            Origin = "https://localhost",
+            RPID = "localhost",
+            Session = response.Session,
+            Response = new
+            {
+                Id = "LcVLKA2QkfwzvuSTxIIyFVTJ9IopE57xTYvJ_0Nx9nk",
+                RawId = "LcVLKA2QkfwzvuSTxIIyFVTJ9IopE57xTYvJ_0Nx9nk",
+                Response = new
+                {
+                    AuthenticatorData = "8egiesVpgMwlnLcY0N3ldtvrzKGUPb763GkSYC4CzTkFAAAAAA",
+                    ClientDataJson =
+                        "eyJ0eXBlIjoid2ViYXV0aG4uZ2V0IiwiY2hhbGxlbmdlIjoiYmhPUllmRlR5S1hfdEtDQkFQbVVKdyIsIm9yaWdpbiI6Imh0dHBzOi8vYWRtaW4ucGFzc3dvcmRsZXNzLmRldiIsImNyb3NzT3JpZ2luIjpmYWxzZX0",
+                    Signature =
+                        "MEUCIQDiTSGMfKb_qZQDB0J8KFFZeAOrvCZEF2yi6MgoUNwkTgIgHYfe8LKPg-INMK9NxJfPaCdNsRUtP2DMhDKraJAOvzk"
+                },
+                type = "public-key"
+            }
+        };
+
+        var result = await _httpClient.PostAsJsonAsync("/signin/complete", payload);
+        var body = await result.Content.ReadAsStringAsync();
+
+        Assert.Equal(HttpStatusCode.BadRequest, result.StatusCode);
+        AssertHelper.AssertEqualJson("""
+                                     {
+                                       "type": "https://docs.passwordless.dev/guide/errors.html#unknown_credential",
+                                       "title": "We don't recognize the passkey you sent us.",
+                                       "status": 400,
+                                       "credentialId": "LcVLKA2QkfwzvuSTxIIyFVTJ9IopE57xTYvJ_0Nx9nk",
+                                       "errorCode": "unknown_credential"
+                                     }
+                                     """, body);
     }
 }
