+++ tests/Api.IntegrationTests/PublicSignInTests.cs
@@ -1,7 +1,5 @@
-﻿using System.Net;
-using System.Net.Http.Json;
+﻿using System.Net.Http.Json;
 using System.Text.Json;
-using Fido2NetLib;
 using Fido2NetLib.Objects;
 using Microsoft.Extensions.DependencyInjection;
 using Passwordless.Service.Models;
@@ -16,49 +14,6 @@ public class PublicSignInTests : PublicTests
     {
     }
 
-    [Fact]
-    public async Task UnknownCredentialThrows()
-    {
-        var options = await PostAsync("/signin/begin", new { Origin = "https://localhost", RPID = "localhost" });
-
-        var response = await options.Content.ReadFromJsonAsync<SessionResponse<AssertionOptions>>();
-
-        var payload = new
-        {
-            Origin = "https://localhost",
-            RPID = "localhost",
-            Session = response.Session,
-            Response = new
-            {
-                Id = "LcVLKA2QkfwzvuSTxIIyFVTJ9IopE57xTYvJ_0Nx9nk",
-                RawId = "LcVLKA2QkfwzvuSTxIIyFVTJ9IopE57xTYvJ_0Nx9nk",
-                Response = new
-                {
-                    AuthenticatorData = "8egiesVpgMwlnLcY0N3ldtvrzKGUPb763GkSYC4CzTkFAAAAAA",
-                    ClientDataJson =
-                        "eyJ0eXBlIjoid2ViYXV0aG4uZ2V0IiwiY2hhbGxlbmdlIjoiYmhPUllmRlR5S1hfdEtDQkFQbVVKdyIsIm9yaWdpbiI6Imh0dHBzOi8vYWRtaW4ucGFzc3dvcmRsZXNzLmRldiIsImNyb3NzT3JpZ2luIjpmYWxzZX0",
-                    Signature =
-                        "MEUCIQDiTSGMfKb_qZQDB0J8KFFZeAOrvCZEF2yi6MgoUNwkTgIgHYfe8LKPg-INMK9NxJfPaCdNsRUtP2DMhDKraJAOvzk"
-                },
-                type = "public-key"
-            }
-        };
-
-        var result = await PostAsync("/signin/complete", payload);
-        var body = await result.Content.ReadAsStringAsync();
-
-        Assert.Equal(HttpStatusCode.BadRequest, result.StatusCode);
-        AssertHelper.AssertEqualJson("""
-        {
-          "type": "https://docs.passwordless.dev/guide/errors.html#unknown_credential",
-          "title": "We don't recognize the passkey you sent us.",
-          "status": 400,          
-          "credentialId": "LcVLKA2QkfwzvuSTxIIyFVTJ9IopE57xTYvJ_0Nx9nk",
-          "errorCode": "unknown_credential"
-        }
-        """, body);
-
-    }
 
     [Fact]
     public async Task CredentialsList()
